backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pandas
scikit-learn
anthropic
twilio
python-dotenv
